+++ NEWS.md
@@ -1,4 +1,4 @@
-# STMR 0.1.6.9000
+# STMR 0.1.7
 
 * Added `conservative` option when generating progression tables, which is equal to `mfactor = 3`
 * In function `plot_progression_table()`, renamed `adjustment_multiplier` to `multiplier`, due to R "error" of assuming user-provided `adjustment` parameter to be `adjustment_multiplier`.